requirements.txt
# Hybrid Ancestry Inference Pipeline - Requirements
# =================================================

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Genomics and bioinformatics
pysam>=0.19.0
cyvcf2>=0.30.0
scikit-allel>=1.3.0

# File I/O and data processing
h5py>=3.1.0
pyyaml>=6.0
requests>=2.25.0
urllib3>=1.26.0

# Statistical analysis
statsmodels>=0.13.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
psutil>=5.8.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Optional: Jupyter for interactive analysis
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Advanced visualization
plotly>=5.0.0